+++ __tests__/branch.test.js
@@ -1,5 +1,5 @@
-const integration = "kochava";
-const name = "Kochava";
+const integration = "branch";
+const name = "Branch";
 
 const fs = require("fs");
 const path = require("path");
@@ -17,12 +17,8 @@ test(`${name} Tests`, async () => {
   const expectedData = JSON.parse(outputDataFile);
   await Promise.all(
     inputData.map(async (input, index) => {
-      try {
-        const output = await transformer.process(input);
-        expect(output).toEqual(expectedData[index]);
-      } catch (error) {
-        expect(error.message).toEqual("message type not supported");
-      }
+      const output = await transformer.process(input);
+      expect(output).toEqual(expectedData[index]);
     })
   );
 });