+++ __tests__/heap.test.js
@@ -1,9 +1,8 @@
-const integration = "kochava";
-const name = "Kochava";
+const integration = "heap";
+const name = "Heap";
 
 const fs = require("fs");
 const path = require("path");
-
 const transformer = require(`../v0/${integration}/transform`);
 
 test(`${name} Tests`, async () => {
@@ -17,12 +16,8 @@ test(`${name} Tests`, async () => {
   const expectedData = JSON.parse(outputDataFile);
   await Promise.all(
     inputData.map(async (input, index) => {
-      try {
-        const output = await transformer.process(input);
-        expect(output).toEqual(expectedData[index]);
-      } catch (error) {
-        expect(error.message).toEqual("message type not supported");
-      }
+      const output = await transformer.process(input);
+      expect(output).toEqual(expectedData[index]);
     })
   );
 });