+++ __tests__/salesforce.test.js
@@ -1,11 +1,11 @@
-const integration = "kochava";
-const name = "Kochava";
+const integration = "salesforce";
+const name = "Salesforce";
 
 const fs = require("fs");
 const path = require("path");
-
 const transformer = require(`../v0/${integration}/transform`);
-
+// const { compareJSON } = require("./util");
+jest.setTimeout(30000);
 test(`${name} Tests`, async () => {
   const inputDataFile = fs.readFileSync(
     path.resolve(__dirname, `./data/${integration}_input.json`)
@@ -15,14 +15,12 @@ test(`${name} Tests`, async () => {
   );
   const inputData = JSON.parse(inputDataFile);
   const expectedData = JSON.parse(outputDataFile);
-  await Promise.all(
-    inputData.map(async (input, index) => {
-      try {
-        const output = await transformer.process(input);
-        expect(output).toEqual(expectedData[index]);
-      } catch (error) {
-        expect(error.message).toEqual("message type not supported");
-      }
-    })
-  );
+
+  inputData.forEach(async (input, index) => {
+    var output = transformer.process(input); 
+    output.header.Authorization = ''
+    expectedData.header.Authorization='' 
+    expect(output).toEqual(expectedData[index]);
+  });
+
 });