+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.2] - 2020-02-07
+
+### Added
+
+- Branch integration 
+- Kochava integration
+- Keen integration
+- Salesforce integration
+- Kissmetrics integration
+- Adjust integration
+
 ## [0.1.1] - 2019-11-05
 
 ### Added